Implement the main mechanics of the game.
Wrote the function for computer choice using
Math.random() function.

Wrote the function for human choice that prompt
users to enter a string value.

Wrote the function playRound which compares both
choices and decides on a winner.

Wrote the function playGame which simulates 5
games, increment each players' scores, and
announces the winner at the end.

The game is played through the Google's console.